* Update build for win target. * Improve tarball build script

+++ packaging/makeTarball.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 
 #Reconfigure
-make distclean
 ./configure
+make distclean
+
+if [ $? -ne 0 ] ; then
+	echo "Something went wrong with distclean. Cannot continue"
+	exit 1
+fi
+
+./configure
+if [ $? -ne 0 ] ; then
+	echo "Something went wrong with configure. Cannot continue"
+	exit 1
+fi
 rm -rf autom4te.cache
 
 #Build tarball
@@ -31,13 +42,52 @@ mkdir tarball
 cd tarball/
 tar -zxf ../3depict-$VER.tar.gz
 rm ../3depict-$VER.tar.gz
-
 #Autoconf buggers up the name case
 mv 3depict-$VER 3Depict-$VER
 
+#--- Translation stuff--
+
+#Build the translation files
+pushd ../translations
+./makeTranslations 
+mkdir -p ../tarball/3Depict-$VER/translations
+cp 3Depict*.mo ../tarball/3Depict-$VER/translations/
+popd
+
+
+pushd 3Depict-$VER/translations/
+
+#Try to locate the corresponding wxMO file, then install it
+for i in `ls *.mo`
+do
+	LOCALEVAL=`basename $i | sed 's/\.mo//' | sed 's/3Depict_//'`
+	WXMO=`find /usr/share/locale/$LOCALEVAL -name wxstd.mo`
+		
+	mkdir -p LC_MESSAGES/$LOCALEVAL
+	mv $i LC_MESSAGES/$LOCALEVAL/3Depict.mo
+
+	#Did we find a wxstd.mo for this locale?
+	if [ x"$WXMO" == x"" ] ; then
+		echo "WARNING: No wxstd.mo found for locale: $LOCALEVAL "
+	else
+		echo  cp $WXMO LC_MESSAGES/$LOCALEVAL/
+	fi
+done
+
+if [ -d LC_MESSAGES ] ; then
+	mv LC_MESSAGES ../
+else
+	echo "ERROR: No Locales built.. Aborting"
+	exit 1
+fi
+
+popd
+
+#--------------
+
+
+
+#make the final source tarball
 tar -cz 3Depict-$VER > 3Depict-$VER.tar.gz
-
-
-
